claude-scripts/requirements.txt
anthropic==0.5.0
python-dotenv==1.0.0
